Script is updated according to comments from first review iteration
